+++ flowmachine/flowmachine/core/server/query_schemas/aggregate_network_objects.py
@@ -9,7 +9,7 @@ from marshmallow_oneofschema import OneOfSchema
 from flowmachine.features import AggregateNetworkObjects
 from .base_exposed_query import BaseExposedQuery
 from .total_network_objects import TotalNetworkObjectsSchema, TotalNetworkObjectsExposed
-from .custom_fields import AggregationUnit, Statistic, AggregateBy
+from .custom_fields import Statistic, AggregateBy
 
 __all__ = ["AggregateNetworkObjectsSchema", "AggregateNetworkObjectsExposed"]
 
@@ -20,6 +20,7 @@ class InputToAggregateNetworkObjectsSchema(OneOfSchema):
 
 
 class AggregateNetworkObjectsSchema(Schema):
+    # query_kind parameter is required here for claims validation
     query_kind = fields.String(validate=OneOf(["aggregate_network_objects"]))
     total_network_objects = fields.Nested(
         InputToAggregateNetworkObjectsSchema, required=True