+++ flowmachine/flowmachine/core/server/query_schemas/unique_subscriber_counts.py
@@ -7,12 +7,13 @@ from marshmallow.validate import OneOf, Length
 
 from flowmachine.features import UniqueSubscriberCounts
 from .base_exposed_query import BaseExposedQuery
-from .custom_fields import AggregationUnit
+from .aggregation_unit import AggregationUnit, get_spatial_unit_obj
 
 __all__ = ["UniqueSubscriberCountsSchema", "UniqueSubscriberCountsExposed"]
 
 
 class UniqueSubscriberCountsSchema(Schema):
+    # query_kind parameter is required here for claims validation
     query_kind = fields.String(validate=OneOf(["unique_subscriber_counts"]))
     start_date = fields.Date(required=True)
     end_date = fields.Date(required=True)
@@ -41,5 +42,7 @@ class UniqueSubscriberCountsExposed(BaseExposedQuery):
         Query
         """
         return UniqueSubscriberCounts(
-            start=self.start_date, stop=self.end_date, level=self.aggregation_unit
+            start=self.start_date,
+            stop=self.end_date,
+            spatial_unit=get_spatial_unit_obj(self.aggregation_unit),
         )