+++ flowmachine/flowmachine/features/location/unique_subscriber_counts.py
@@ -4,7 +4,7 @@
 
 # -*- coding: utf-8 -*-
 
-from typing import List
+from typing import List, Union
 
 """
 Class for UniqueSubscriberCounts. UniqueSubscriberCounts counts
@@ -17,16 +17,17 @@ visited in a given period of time.
 """
 from ...core.query import Query
 from ...core.mixins import GeoDataMixin
+from ...core import make_spatial_unit
+from ...core.spatial_unit import AnySpatialUnit
 
-from flowmachine.utils import get_columns_for_level
-from ..utilities.subscriber_locations import subscriber_locations
+from ..utilities.subscriber_locations import SubscriberLocations
 
 
 class UniqueSubscriberCounts(GeoDataMixin, Query):
 
     """
     Class that defines counts of unique subscribers for each location.
-    Each location for the given level is accompanied by the count of unique subscribers.
+    Each location for the given spatial unit is accompanied by the count of unique subscribers.
 
     Parameters
     ----------
@@ -35,29 +36,9 @@ class UniqueSubscriberCounts(GeoDataMixin, Query):
         e.g. 2016-01-01 or 2016-01-01 14:03:01
     stop : str
         As above
-    level : str, default 'cell'
-        Levels can be one of:
-            'cell':
-                The identifier as it is found in the CDR itself
-            'versioned-cell':
-                The identifier as found in the CDR combined with the version from
-                the cells table.
-            'versioned-site':
-                The ID found in the sites table, coupled with the version
-                number.
-            'polygon':
-                A custom set of polygons that live in the database. In which
-                case you can pass the parameters column_name, which is the column
-                you want to return after the join, and table_name, the table where
-                the polygons reside (with the schema), and additionally geom_col
-                which is the column with the geometry information (will default to
-                'geom')
-            'admin*':
-                An admin region of interest, such as admin3. Must live in the
-                database in the standard location.
-            'grid':
-                A square in a regular grid, in addition pass size to
-                determine the size of the polygon.
+    spatial_unit : flowmachine.core.spatial_unit.*SpatialUnit, default cell
+        Spatial unit to which subscriber locations will be mapped. See the
+        docstring of make_spatial_unit for more information.
     hours : tuple of ints, default 'all'
         subset the result within certain hours, e.g. (4,17)
         This will subset the query only with these hours, but
@@ -79,16 +60,10 @@ class UniqueSubscriberCounts(GeoDataMixin, Query):
         these lines with null cells should still be present, although they contain
         no information on the subscribers location, they still tell us that the subscriber made
         a call at that time.
-    column_name : str, optional
-        Option, none-standard, name of the column that identifies the
-        spatial level, i.e. could pass admin3pcod to use the admin 3 pcode
-        as opposed to the name of the region.
-    kwargs :
-        Eventually passed to flowmachine.JoinToLocation.
 
     Examples
     --------
-    >>> usc = UniqueSubscriberCounts('2016-01-01', '2016-01-04', level = 'admin3', hours = (5,17))
+    >>> usc = UniqueSubscriberCounts('2016-01-01', '2016-01-04', spatial_unit=admin_spatial_unit(level=3), hours=(5,17))
     >>> usc.head(4)
           name                  unique_subscriber_counts
     0     Arghakhanchi          313
@@ -100,11 +75,9 @@ class UniqueSubscriberCounts(GeoDataMixin, Query):
         self,
         start,
         stop,
-        level="cell",
+        spatial_unit: AnySpatialUnit = make_spatial_unit("cell"),
         hours="all",
         table="all",
-        column_name=None,
-        **kwargs
     ):
         """
 
@@ -112,28 +85,22 @@ class UniqueSubscriberCounts(GeoDataMixin, Query):
 
         self.start = start
         self.stop = stop
-        self.level = level
+        self.spatial_unit = spatial_unit
         self.hours = hours
         self.table = table
-        self.column_name = column_name
-        self._kwargs = kwargs
-        self.ul = subscriber_locations(
+        self.ul = SubscriberLocations(
             start=self.start,
             stop=self.stop,
-            level=self.level,
+            spatial_unit=self.spatial_unit,
             hours=self.hours,
             table=self.table,
-            column_name=self.column_name,
-            **kwargs
         )
 
         super().__init__()
 
     @property
     def column_names(self) -> List[str]:
-        return get_columns_for_level(self.level, self.column_name) + [
-            "unique_subscriber_counts"
-        ]
+        return self.spatial_unit.location_id_columns + ["unique_subscriber_counts"]
 
     def _make_query(self):
         """
@@ -141,7 +108,7 @@ class UniqueSubscriberCounts(GeoDataMixin, Query):
         metaclass Query().
         """
 
-        relevant_columns = ",".join(get_columns_for_level(self.level, self.column_name))
+        relevant_columns = ",".join(self.spatial_unit.location_id_columns)
         sql = """
         SELECT {rc}, COUNT(unique_subscribers) AS unique_subscriber_counts FROM 
         (SELECT